scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.2.0
boto3>=1.34.0
